+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Warn the user if a new CFG is being created, when the project has been moved/renamed.
 * Some Python cleaning about handling file path, related to CFG file for instance
 
 ## 3.11.1 - 2023-03-21